Consolidate public API under statscache.api
Move all modules that are intended to be used by plugins into the statscache.api
subpackage.
