+++ statscache/app.py
@@ -1,6 +1,5 @@
 import flask
 import fedmsg.config
-import statscache.plugins
 import statscache.utils
 import statscache.frequency
 import copy
@@ -20,7 +19,7 @@ plugins = {
 } # mapping of identifiers to plugin instances
 
 uri = config['statscache.sqlalchemy.uri']
-session = statscache.plugins.init_model(uri)
+session = statscache.utils.init_model(uri)
 
 
 def wants_pagination():
